requirements.txt
fastapi
uvicorn[standard]
opencc-python-reimplemented
translators
pycryptodome
markdownify
httpx